+++ extensions/name-com/src/check-account-balance.ts
@@ -1,13 +1,10 @@
-import { API_URL, headers, parseResponse } from "./api";
+import { callCoreApi } from "./api";
 import { showToast, Toast, updateCommandMetadata } from "@raycast/api";
 
 export default async function CheckAccountBalance() {
   const toast = await showToast(Toast.Style.Animated, "Checking");
   try {
-    const response = await fetch(API_URL + "accountinfo/balance", {
-      headers,
-    });
-    const result = await parseResponse(response) as { balance: number };
+    const result = await callCoreApi<{ balance: number }>("accountinfo/balance");
     const strBalance = `$${result.balance}`;
     await updateCommandMetadata({ subtitle: `Name.com - Balance: ${strBalance}` });
     toast.style = Toast.Style.Success;