+++ extensions/name-com/src/tools/list-dns-records.ts
@@ -4,8 +4,8 @@ import { DNSRecord } from "../types";
 type Input = {
   /**
    * The domain name
-  */
- domainName: string;
+   */
+  domainName: string;
 };
 
 /**